`tmp_dir` and `create_tmp_file` fixtures moved to `fixtures.py`.
